split types into different files
