+++ server/statistics/hot_cache.go
@@ -16,9 +16,7 @@ package statistics
 import (
 	"fmt"
 	"math/rand"
-	"time"
 
-	"github.com/pingcap/kvproto/pkg/metapb"
 	"github.com/pingcap/pd/pkg/cache"
 	"github.com/pingcap/pd/server/core"
 )
@@ -41,260 +39,10 @@ const (
 	hotRegionAntiCount         = 1
 )
 
-// FlowKind is a identify Flow types.
-type FlowKind uint32
-
-// Flags for flow.
-const (
-	WriteFlow FlowKind = iota
-	ReadFlow
-)
-
-func (k FlowKind) String() string {
-	switch k {
-	case WriteFlow:
-		return "write"
-	case ReadFlow:
-		return "read"
-	}
-	return "unimplemented"
-}
-
-// HotStoresStats saves the hotspot peer's statistics.
-type HotStoresStats struct {
-	hotStoreStats  map[uint64]cache.Cache         // storeID -> hot regions
-	storesOfRegion map[uint64]map[uint64]struct{} // regionID -> storeIDs
-}
-
-// NewHotStoresStats creates a HotStoresStats
-func NewHotStoresStats() *HotStoresStats {
-	return &HotStoresStats{
-		hotStoreStats:  make(map[uint64]cache.Cache),
-		storesOfRegion: make(map[uint64]map[uint64]struct{}),
-	}
-}
-
-// CheckRegionFlow checks the flow information of region.
-func (f *HotStoresStats) CheckRegionFlow(region *core.RegionInfo, kind FlowKind) []HotSpotPeerStatGenerator {
-	var (
-		generators   []HotSpotPeerStatGenerator
-		getBytesFlow func() uint64
-		getKeysFlow  func() uint64
-		bytesPerSec  uint64
-		keysPerSec   uint64
-
-		isExpiredInStore func(region *core.RegionInfo, storeID uint64) bool
-	)
-
-	storeIDs := make(map[uint64]struct{})
-	// gets the storeIDs, including old region and new region
-	ids, ok := f.storesOfRegion[region.GetID()]
-	if ok {
-		for storeID := range ids {
-			storeIDs[storeID] = struct{}{}
-		}
-	}
-
-	for _, peer := range region.GetPeers() {
-		// ReadFlow no need consider the followers.
-		if kind == ReadFlow && peer.GetStoreId() != region.GetLeader().GetStoreId() {
-			continue
-		}
-		if _, ok := storeIDs[peer.GetStoreId()]; !ok {
-			storeIDs[peer.GetStoreId()] = struct{}{}
-		}
-	}
-
-	switch kind {
-	case WriteFlow:
-		getBytesFlow = region.GetBytesWritten
-		getKeysFlow = region.GetKeysWritten
-		isExpiredInStore = func(region *core.RegionInfo, storeID uint64) bool {
-			return region.GetStorePeer(storeID) == nil
-		}
-	case ReadFlow:
-		getBytesFlow = region.GetBytesRead
-		getKeysFlow = region.GetKeysRead
-		isExpiredInStore = func(region *core.RegionInfo, storeID uint64) bool {
-			return region.GetLeader().GetStoreId() != storeID
-		}
-	}
-
-	bytesPerSecInit := uint64(float64(getBytesFlow()) / float64(RegionHeartBeatReportInterval))
-	keysPerSecInit := uint64(float64(getKeysFlow()) / float64(RegionHeartBeatReportInterval))
-	for storeID := range storeIDs {
-		bytesPerSec = bytesPerSecInit
-		keysPerSec = keysPerSecInit
-		var oldRegionStat *HotSpotPeerStat
-
-		hotStoreStats, ok := f.hotStoreStats[storeID]
-		if ok {
-			if v, isExist := hotStoreStats.Peek(region.GetID()); isExist {
-				oldRegionStat = v.(*HotSpotPeerStat)
-				// This is used for the simulator.
-				if Denoising {
-					interval := time.Since(oldRegionStat.LastUpdateTime).Seconds()
-					if interval < minHotRegionReportInterval && !isExpiredInStore(region, storeID) {
-						continue
-					}
-					bytesPerSec = uint64(float64(getBytesFlow()) / interval)
-					keysPerSec = uint64(float64(getKeysFlow()) / interval)
-				}
-			}
-		}
-
-		generator := &hotSpotPeerStatGenerator{
-			Region:    region,
-			StoreID:   storeID,
-			FlowBytes: bytesPerSec,
-			FlowKeys:  keysPerSec,
-			Kind:      kind,
-
-			lastHotSpotPeerStats: oldRegionStat,
-		}
-
-		if isExpiredInStore(region, storeID) {
-			generator.Expired = true
-		}
-		generators = append(generators, generator)
-	}
-	return generators
-}
-
-// Update updates the items in statistics.
-func (f *HotStoresStats) Update(item *HotSpotPeerStat) {
-	if item.IsNeedDelete() {
-		if hotStoreStat, ok := f.hotStoreStats[item.StoreID]; ok {
-			hotStoreStat.Remove(item.RegionID)
-		}
-		if index, ok := f.storesOfRegion[item.RegionID]; ok {
-			delete(index, item.StoreID)
-		}
-	} else {
-		hotStoreStat, ok := f.hotStoreStats[item.StoreID]
-		if !ok {
-			hotStoreStat = cache.NewCache(statCacheMaxLen, cache.TwoQueueCache)
-			f.hotStoreStats[item.StoreID] = hotStoreStat
-		}
-		hotStoreStat.Put(item.RegionID, item)
-		index, ok := f.storesOfRegion[item.RegionID]
-		if !ok {
-			index = make(map[uint64]struct{})
-		}
-		index[item.StoreID] = struct{}{}
-		f.storesOfRegion[item.RegionID] = index
-	}
-}
-
-func (f *HotStoresStats) isRegionHotWithAnyPeers(region *core.RegionInfo, hotThreshold int) bool {
-	for _, peer := range region.GetPeers() {
-		if f.isRegionHotWithPeer(region, peer, hotThreshold) {
-			return true
-		}
-	}
-	return false
-
-}
-
-func (f *HotStoresStats) isRegionHotWithPeer(region *core.RegionInfo, peer *metapb.Peer, hotThreshold int) bool {
-	if peer == nil {
-		return false
-	}
-	storeID := peer.GetStoreId()
-	stats, ok := f.hotStoreStats[storeID]
-	if !ok {
-		return false
-	}
-	if stat, ok := stats.Peek(region.GetID()); ok {
-		return stat.(*HotSpotPeerStat).HotDegree >= hotThreshold
-	}
-	return false
-}
-
-// HotSpotPeerStatGenerator used to produce new hotspot statistics.
-type HotSpotPeerStatGenerator interface {
-	GenHotSpotPeerStats(stats *StoresStats) *HotSpotPeerStat
-}
-
-// hotSpotPeerStatBuilder used to produce new hotspot statistics.
-type hotSpotPeerStatGenerator struct {
-	Region    *core.RegionInfo
-	StoreID   uint64
-	FlowKeys  uint64
-	FlowBytes uint64
-	Expired   bool
-	Kind      FlowKind
-
-	lastHotSpotPeerStats *HotSpotPeerStat
-}
-
-const rollingWindowsSize = 5
-
-// GenHotSpotPeerStats implements HotSpotPeerStatsGenerator.
-func (flowStats *hotSpotPeerStatGenerator) GenHotSpotPeerStats(stats *StoresStats) *HotSpotPeerStat {
-	var hotRegionThreshold uint64
-	switch flowStats.Kind {
-	case WriteFlow:
-		hotRegionThreshold = calculateWriteHotThresholdWithStore(stats, flowStats.StoreID)
-	case ReadFlow:
-		hotRegionThreshold = calculateReadHotThresholdWithStore(stats, flowStats.StoreID)
-	}
-	flowBytes := flowStats.FlowBytes
-	oldItem := flowStats.lastHotSpotPeerStats
-	region := flowStats.Region
-	newItem := &HotSpotPeerStat{
-		RegionID:       region.GetID(),
-		FlowBytes:      flowStats.FlowBytes,
-		FlowKeys:       flowStats.FlowKeys,
-		LastUpdateTime: time.Now(),
-		StoreID:        flowStats.StoreID,
-		Version:        region.GetMeta().GetRegionEpoch().GetVersion(),
-		AntiCount:      hotRegionAntiCount,
-		Kind:           flowStats.Kind,
-		needDelete:     flowStats.Expired,
-	}
-
-	if region.GetLeader().GetStoreId() == flowStats.StoreID {
-		newItem.isLeader = true
-	}
-
-	if newItem.IsNeedDelete() {
-		return newItem
-	}
-
-	if oldItem != nil {
-		newItem.HotDegree = oldItem.HotDegree + 1
-		newItem.Stats = oldItem.Stats
-	}
-
-	if flowBytes >= hotRegionThreshold {
-		if oldItem == nil {
-			newItem.Stats = NewRollingStats(rollingWindowsSize)
-		}
-		newItem.isNew = true
-		newItem.Stats.Add(float64(flowBytes))
-		return newItem
-	}
-
-	// smaller than hotRegionThreshold
-	if oldItem == nil {
-		return nil
-	}
-	if oldItem.AntiCount <= 0 {
-		newItem.needDelete = true
-		return newItem
-	}
-	// eliminate some noise
-	newItem.HotDegree = oldItem.HotDegree - 1
-	newItem.AntiCount = oldItem.AntiCount - 1
-	newItem.Stats.Add(float64(flowBytes))
-	return newItem
-}
-
 // HotSpotCache is a cache hold hot regions.
 type HotSpotCache struct {
-	writeFlow *HotStoresStats
-	readFlow  *HotStoresStats
+	writeFlow *hotStoresStats
+	readFlow  *hotStoresStats
 }
 
 // NewHotSpotCache creates a new hot spot cache.
@@ -331,19 +79,9 @@ func (w *HotSpotCache) CheckRead(region *core.RegionInfo, stats *StoresStats) []
 	return updateItems
 }
 
-func (w *HotSpotCache) incMetrics(name string, storeID uint64, kind FlowKind) {
-	storeTag := fmt.Sprintf("store-%d", storeID)
-	switch kind {
-	case WriteFlow:
-		hotCacheStatusGauge.WithLabelValues(name, storeTag, "write").Inc()
-	case ReadFlow:
-		hotCacheStatusGauge.WithLabelValues(name, storeTag, "read").Inc()
-	}
-}
-
 // Update updates the cache.
 func (w *HotSpotCache) Update(item *HotSpotPeerStat) {
-	var stats *HotStoresStats
+	var stats *hotStoresStats
 	switch item.Kind {
 	case WriteFlow:
 		stats = w.writeFlow
@@ -398,6 +136,16 @@ func (w *HotSpotCache) RandHotRegionFromStore(storeID uint64, kind FlowKind, hot
 	return nil
 }
 
+// IsRegionHot checks if the region is hot.
+func (w *HotSpotCache) IsRegionHot(region *core.RegionInfo, hotThreshold int) bool {
+	stats := w.writeFlow
+	if stats.isRegionHotWithAnyPeers(region, hotThreshold) {
+		return true
+	}
+	stats = w.readFlow
+	return stats.isRegionHotWithPeer(region, region.GetLeader(), hotThreshold)
+}
+
 // CollectMetrics collect the hot cache metrics
 func (w *HotSpotCache) CollectMetrics(stats *StoresStats) {
 	for storeID, flowStats := range w.writeFlow.hotStoreStats {
@@ -415,14 +163,14 @@ func (w *HotSpotCache) CollectMetrics(stats *StoresStats) {
 	}
 }
 
-// IsRegionHot checks if the region is hot.
-func (w *HotSpotCache) IsRegionHot(region *core.RegionInfo, hotThreshold int) bool {
-	stats := w.writeFlow
-	if stats.isRegionHotWithAnyPeers(region, hotThreshold) {
-		return true
+func (w *HotSpotCache) incMetrics(name string, storeID uint64, kind FlowKind) {
+	storeTag := fmt.Sprintf("store-%d", storeID)
+	switch kind {
+	case WriteFlow:
+		hotCacheStatusGauge.WithLabelValues(name, storeTag, "write").Inc()
+	case ReadFlow:
+		hotCacheStatusGauge.WithLabelValues(name, storeTag, "read").Inc()
 	}
-	stats = w.readFlow
-	return stats.isRegionHotWithPeer(region, region.GetLeader(), hotThreshold)
 }
 
 // Utils
@@ -446,12 +194,4 @@ func calculateReadHotThresholdWithStore(stats *StoresStats, storeID uint64) uint
 		hotRegionThreshold = hotReadRegionMinFlowRate
 	}
 	return hotRegionThreshold
-}
-
-// RegionStatInformer provides access to a shared informer of statistics.
-type RegionStatInformer interface {
-	IsRegionHot(region *core.RegionInfo) bool
-	RegionWriteStats() map[uint64][]*HotSpotPeerStat
-	RegionReadStats() map[uint64][]*HotSpotPeerStat
-	RandHotRegionFromStore(store uint64, kind FlowKind) *core.RegionInfo
 }
