+++ server/statistics/region.go
@@ -14,63 +14,8 @@
 package statistics
 
 import (
-	"time"
-
 	"github.com/pingcap/pd/server/core"
 )
-
-// HotSpotPeerStat records each hot region's statistics
-type HotSpotPeerStat struct {
-	RegionID  uint64 `json:"region_id"`
-	FlowBytes uint64 `json:"flow_bytes"`
-	FlowKeys  uint64 `json:"flow_keys"`
-	// HotDegree records the hot region update times
-	HotDegree int `json:"hot_degree"`
-	// LastUpdateTime used to calculate average write
-	LastUpdateTime time.Time `json:"last_update_time"`
-	// StoreID is the store id of the region peer
-	StoreID uint64   `json:"store_id"`
-	Kind    FlowKind `json:"kind"`
-	// AntiCount used to eliminate some noise when remove region in cache
-	AntiCount int
-	// Version used to check the region split times
-	Version uint64
-	// Stats is a rolling statistics, recording some recently added records.
-	Stats *RollingStats
-
-	needDelete bool
-	isLeader   bool
-	isNew      bool
-}
-
-// IsNeedDelete to delete the item in cache.
-func (stat HotSpotPeerStat) IsNeedDelete() bool {
-	return stat.needDelete
-}
-
-// IsLeader indicaes the item belong to the leader.
-func (stat HotSpotPeerStat) IsLeader() bool {
-	return stat.isLeader
-}
-
-// IsNew indicaes the item is first update in the cache of the region.
-func (stat HotSpotPeerStat) IsNew() bool {
-	return stat.isNew
-}
-
-// RegionsStat is a list of a group region state type
-type RegionsStat []HotSpotPeerStat
-
-func (m RegionsStat) Len() int           { return len(m) }
-func (m RegionsStat) Swap(i, j int)      { m[i], m[j] = m[j], m[i] }
-func (m RegionsStat) Less(i, j int) bool { return m[i].FlowBytes < m[j].FlowBytes }
-
-// HotRegionsStat records all hot regions statistics
-type HotRegionsStat struct {
-	TotalFlowBytes uint64      `json:"total_flow_bytes"`
-	RegionsCount   int         `json:"regions_count"`
-	RegionsStat    RegionsStat `json:"statistics"`
-}
 
 // RegionStats records a list of regions' statistics and distribution status.
 type RegionStats struct {
@@ -84,6 +29,17 @@ type RegionStats struct {
 	StoreLeaderKeys  map[uint64]int64 `json:"store_leader_keys"`
 	StorePeerSize    map[uint64]int64 `json:"store_peer_size"`
 	StorePeerKeys    map[uint64]int64 `json:"store_peer_keys"`
+}
+
+// GetRegionStats scans regions that intersect range [startKey, endKey) and sums up
+// their statistics.
+func GetRegionStats(f func(startKey, endKey []byte) []*core.RegionInfo, startKey, endKey []byte) *RegionStats {
+	stats := newRegionStats()
+	regions := f(startKey, endKey)
+	for _, region := range regions {
+		stats.Observe(region)
+	}
+	return stats
 }
 
 func newRegionStats() *RegionStats {
@@ -121,15 +77,4 @@ func (s *RegionStats) Observe(r *core.RegionInfo) {
 		s.StorePeerSize[storeID] += approximateSize
 		s.StorePeerKeys[storeID] += approximateKeys
 	}
-}
-
-// GetRegionStats scans regions that intersect range [startKey, endKey) and sums up
-// their statistics.
-func GetRegionStats(f func(startKey, endKey []byte) []*core.RegionInfo, startKey, endKey []byte) *RegionStats {
-	stats := newRegionStats()
-	regions := f(startKey, endKey)
-	for _, region := range regions {
-		stats.Observe(region)
-	}
-	return stats
 }
